+++ src/backend/API/Models/ItemModels.cs
@@ -71,9 +71,9 @@ namespace API.Models
 
         [Required(ErrorMessage = "Grup ID zorunludur")]
         public int GroupId { get; set; }
-        public int? X { get; set; }
-        public int? Y { get; set; }
-        public int? Z { get; set; }
+        public double? X { get; set; }
+        public double? Y { get; set; }
+        public double? Z { get; set; }
 
         [MaxLength(500, ErrorMessage = "Resim URL'i en fazla 500 karakter olabilir")]
         public string? ImageUrl { get; set; }
@@ -102,9 +102,9 @@ namespace API.Models
         [Required(ErrorMessage = "Grup ID zorunludur")]
         public int GroupId { get; set; }
 
-        public int? X { get; set; }
-        public int? Y { get; set; }
-        public int? Z { get; set; }
+        public double? X { get; set; }
+        public double? Y { get; set; }
+        public double? Z { get; set; }
 
         [MaxLength(500, ErrorMessage = "Resim URL'i en fazla 500 karakter olabilir")]
         public string? ImageUrl { get; set; }
@@ -142,9 +142,9 @@ namespace API.Models
         public string DocNumber { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
 
-        public int? X { get; set; }
-        public int? Y { get; set; }
-        public int? Z { get; set; }
+        public double? X { get; set; }
+        public double? Y { get; set; }
+        public double? Z { get; set; }
         public string? ImageUrl { get; set; }
         public bool? Cancelled { get; set; }
         public DateTime CreatedAt { get; set; }
